+++ docs/changelog.md
@@ -15,6 +15,7 @@ pageflow_next_text: Contributing
 * added helpers to standardise the structure of format strings and parameter lists that get fed into `ParameterisedException`
   - added `BuildThrownAndCalledBy`
   - added `BuildThrownBy`
+  - updated `UnsupportedType` to use the new helpers
 
 ## v1.2016061201